+++ instrumentation/excon/lib/opentelemetry/instrumentation/excon/middlewares/old/tracer_middleware.rb
@@ -11,27 +11,17 @@ module OpenTelemetry
         module Old
           # Excon middleware for instrumentation
           class TracerMiddleware < ::Excon::Middleware::Base
-            HTTP_METHODS_TO_UPPERCASE = %w[connect delete get head options patch post put trace].each_with_object({}) do |method, hash|
-              uppercase_method = method.upcase
-              hash[method] = uppercase_method
-              hash[method.to_sym] = uppercase_method
-              hash[uppercase_method] = uppercase_method
-            end.freeze
-
-            HTTP_METHODS_TO_SPAN_NAMES = HTTP_METHODS_TO_UPPERCASE.values.each_with_object({}) do |uppercase_method, hash|
-              hash[uppercase_method] ||= "HTTP #{uppercase_method}"
-            end.freeze
-
             # Constant for the HTTP status range
             HTTP_STATUS_SUCCESS_RANGE = (100..399)
 
             def request_call(datum)
               return @stack.request_call(datum) if untraced?(datum)
 
-              http_method = HTTP_METHODS_TO_UPPERCASE[datum[:method]]
+              span_data = HttpHelper.span_attrs_for(datum[:method], semconv: :old)
+
               attributes = {
                 OpenTelemetry::SemanticConventions::Trace::HTTP_HOST => datum[:host],
-                OpenTelemetry::SemanticConventions::Trace::HTTP_METHOD => http_method,
+                OpenTelemetry::SemanticConventions::Trace::HTTP_METHOD => span_data.normalized_method,
                 OpenTelemetry::SemanticConventions::Trace::HTTP_SCHEME => datum[:scheme],
                 OpenTelemetry::SemanticConventions::Trace::HTTP_TARGET => datum[:path],
                 OpenTelemetry::SemanticConventions::Trace::HTTP_URL => OpenTelemetry::Common::Utilities.cleanse_url(::Excon::Utils.request_uri(datum)),
@@ -41,7 +31,7 @@ module OpenTelemetry
               peer_service = Excon::Instrumentation.instance.config[:peer_service]
               attributes[OpenTelemetry::SemanticConventions::Trace::PEER_SERVICE] = peer_service if peer_service
               attributes.merge!(OpenTelemetry::Common::HTTP::ClientContext.attributes)
-              span = tracer.start_span(HTTP_METHODS_TO_SPAN_NAMES[http_method], attributes: attributes, kind: :client)
+              span = tracer.start_span(span_data.span_name, attributes: attributes, kind: :client)
               ctx = OpenTelemetry::Trace.context_with_span(span)
               datum[:otel_span] = span
               datum[:otel_token] = OpenTelemetry::Context.attach(ctx)